feat: create the repositories classes and move all methods to its corresponding repository
Close: #1
