+++ RestaurantReservation.DB/Repositories/ReservationsRepository.cs
@@ -1,0 +1,60 @@
+using RestaurantReservation.DB.EFContext;
+using RestaurantReservation.DB.Exceptions;
+using RestaurantReservation.DB.Models;
+
+namespace RestaurantReservation.DB.Repositories;
+
+public class ReservationsRepository
+{
+    private readonly RestaurantContext _restaurantContext;
+
+    public ReservationsRepository(RestaurantContext restaurantContext)
+    {
+        _restaurantContext = restaurantContext;
+    }
+    
+    public async Task AddNewReservation(Reservation reservation)
+    {
+        if (!_restaurantContext.Reservations.Any(r => r.ReservationDate == reservation.ReservationDate && r.PartySize == reservation.PartySize))
+        {
+            await _restaurantContext.Reservations.AddAsync(reservation);
+            await _restaurantContext.SaveChangesAsync();
+        }
+        else
+        {
+            throw new RecordExistsException("The reservation already exists in the database!");
+        }
+    }
+
+    public async Task DeleteReservation(Reservation reservation)
+    {
+        if (_restaurantContext.Reservations.Any(r => r.ReservationDate == reservation.ReservationDate && r.PartySize == reservation.PartySize))
+        {
+            _restaurantContext.Reservations.Remove(reservation);
+            await _restaurantContext.SaveChangesAsync();
+        }
+        else
+        {
+            throw new RecordDoesNotExistException("The reservations does not exist in the database!");
+        }
+    }
+
+    public async Task UpdateReservation(Reservation reservation)
+    {
+        var existingReservation = await _restaurantContext.Reservations.FindAsync(reservation.ReservationId);
+        if (existingReservation is not null)
+        {
+            _restaurantContext.Entry(existingReservation).CurrentValues.SetValues(reservation);
+            await _restaurantContext.SaveChangesAsync();
+        }
+        else
+        {
+            throw new RecordDoesNotExistException("The reservations does not exist in the database!");
+        }
+    }
+    
+    public List<Reservation> GetReservationsByCustomer(int customerId)
+    {
+        return _restaurantContext.Reservations.Where(r => r.CustomerId == customerId).ToList();
+    }
+}
