+++ RestaurantReservation/Program.cs
@@ -1,11 +1,18 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using RestaurantReservation.DB.EFContext;
-using RestaurantReservation.DB.Queries;
 using RestaurantReservation.DB.Exceptions;
 using RestaurantReservation.DB.InitialData;
 using RestaurantReservation.DB.Models;
+using RestaurantReservation.DB.Repositories;
 
 var context = new RestaurantContext();
+var ordersRepo = new OrdersRepository(context);
+var customersRepo = new CustomersRepository(context);
+var restaurantsRepo = new RestaurantsRepository(context);
+var employeesRepo = new EmployeeRepository(context);
+var menuItemsRepo = new MenuItemsRepository(context);
+var tablesRepo = new TablesRepository(context);
+var reservationsRepo = new ReservationsRepository(context);
 
 try
 {
@@ -39,13 +46,12 @@ finally
 
 async Task CrudOperateOnEntities(RestaurantContext context)
 {
-    var crudOperate = new EntitiesCrudOperate(context);
-    
+
     #region Customer
 
     // ::: Adding new Entity; Begin.
     
-    await crudOperate.AddNewCustomer(new Customer()
+    await customersRepo.AddNewCustomer(new Customer()
     {
         FirstName = "Mohammed",
         LastName = "Salem",
@@ -59,7 +65,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     // ::: Adding new Entity; End.
 
     // ::: Deleting Entity; Begin.
-    await crudOperate.DeleteCustomer(await context.Customers.FirstAsync(c => c.Email == "test@test.com"));
+    await customersRepo.DeleteCustomer(await context.Customers.FirstAsync(c => c.Email == "test@test.com"));
     var deletedMohSalem = await context.Customers.FirstOrDefaultAsync(c => c.Email == "test@test.com");
     
     Console.WriteLine(deletedMohSalem is null ? "Mohammed Salem got deleted successfully!" : "ERROR while deleting the customer!");
@@ -83,13 +89,13 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         PhoneNumber = "987654321",
     };
     
-    await crudOperate.AddNewCustomer(initialCustomer);
+    await customersRepo.AddNewCustomer(initialCustomer);
     var customerBeforeUpdate = await context.Customers.FirstOrDefaultAsync(c => c.Email == "test@test.com");
     customerToUpdateWith.CustomerId = customerBeforeUpdate!.CustomerId;
     
     Console.WriteLine($"Customer before update: {customerBeforeUpdate?.CustomerId} - {customerBeforeUpdate?.FirstName}, {customerBeforeUpdate?.LastName}.");
 
-    await crudOperate.UpdateCustomer(customerToUpdateWith);
+    await customersRepo.UpdateCustomer(customerToUpdateWith);
     var updatedCustomer = await context.Customers.FirstOrDefaultAsync(c => c.CustomerId == customerToUpdateWith.CustomerId);
     
     Console.WriteLine($"Customer after updating: {updatedCustomer?.CustomerId} - {updatedCustomer?.FirstName}, {updatedCustomer?.LastName}.");
@@ -100,7 +106,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     Console.WriteLine("Customers region clean up!");
     
-    await crudOperate.DeleteCustomer(updatedCustomer!);
+    await customersRepo.DeleteCustomer(updatedCustomer!);
     Console.WriteLine("Customer got deleted successfully!");
 
     #endregion
@@ -109,7 +115,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     // ::: Adding new Entity; Begin.
 
-    await crudOperate.AddNewRestaurant(new Restaurant()
+    await restaurantsRepo.AddNewRestaurant(new Restaurant()
     {
         Name = "Restaurant One",
         Address = "Gaza, Palestine",
@@ -125,7 +131,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
 
     // ::: Deleting Entity; Begin.
     
-    await crudOperate.DeleteRestaurant(await context.Restaurants.FirstAsync(c => c.PhoneNumber == "11223344"));
+    await restaurantsRepo.DeleteRestaurant(await context.Restaurants.FirstAsync(c => c.PhoneNumber == "11223344"));
     var deletedRes = await context.Restaurants.FirstOrDefaultAsync(c => c.PhoneNumber == "11223344");
     
     Console.WriteLine(deletedRes is null ? "Restaurant one got deleted successfully!" : "ERROR while deleting the restaurant!");
@@ -149,13 +155,13 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         OpeningHours = new TimeOnly(),
     };
     
-    await crudOperate.AddNewRestaurant(initialRestaurant);
+    await restaurantsRepo.AddNewRestaurant(initialRestaurant);
     var restaurantBeforeUpdate = await context.Restaurants.FirstOrDefaultAsync(r => r.PhoneNumber == "22334455");
     restaurantToUpdateWith.RestaurantId = restaurantBeforeUpdate!.RestaurantId;
     
     Console.WriteLine($"Restaurant before update: {restaurantBeforeUpdate?.RestaurantId} - {restaurantBeforeUpdate?.Name}, {restaurantBeforeUpdate?.OpeningHours}.");
 
-    await crudOperate.UpdateRestaurant(restaurantToUpdateWith);
+    await restaurantsRepo.UpdateRestaurant(restaurantToUpdateWith);
     var updatedRestaurant = await context.Restaurants.FirstOrDefaultAsync(r => r.RestaurantId == restaurantToUpdateWith.RestaurantId);
     
     Console.WriteLine($"Restaurant after updating: {updatedRestaurant?.RestaurantId} - {updatedRestaurant?.Name}, {updatedRestaurant?.Address}.");
@@ -166,7 +172,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     Console.WriteLine("Restaurants region clean up!");
     
-    await crudOperate.DeleteRestaurant(await context.Restaurants.FirstAsync(c => c.RestaurantId == restaurantToUpdateWith.RestaurantId));
+    await restaurantsRepo.DeleteRestaurant(await context.Restaurants.FirstAsync(c => c.RestaurantId == restaurantToUpdateWith.RestaurantId));
     Console.WriteLine("Restaurant in the restaurants region got deleted successfully!");
 
     #endregion
@@ -175,7 +181,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     // ::: Adding new Entity; Begin.
 
-    await crudOperate.AddNewRestaurant(new Restaurant()
+    await restaurantsRepo.AddNewRestaurant(new Restaurant()
     {
         Name = "Restaurant Three",
         Address = "Gaza, Palestine",
@@ -185,7 +191,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
 
     var restaurantThree = await context.Restaurants.FirstOrDefaultAsync(r => r.PhoneNumber == "999888776");
     
-    await crudOperate.AddNewEmployee(new Employee()
+    await employeesRepo.AddNewEmployee(new Employee()
     {
         FirstName = "Mohammed",
         LastName = "Salem",
@@ -201,7 +207,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     // ::: Deleting Entity; Begin.
     
-    await crudOperate.DeleteEmployee(await context.Employees.FirstAsync(e => e.FirstName == "Mohammed" && e.LastName == "Salem"));
+    await employeesRepo.DeleteEmployee(await context.Employees.FirstAsync(e => e.FirstName == "Mohammed" && e.LastName == "Salem"));
     var deletedEmp = await context.Employees.FirstOrDefaultAsync(e => e.FirstName == "Mohammed" && e.LastName == "Salem");
     
     Console.WriteLine(deletedEmp is null ? "Employee one got deleted successfully!" : "ERROR while deleting the employee!");
@@ -225,13 +231,13 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         RestaurantId = restaurantThree.RestaurantId,
     };
     
-    await crudOperate.AddNewEmployee(initialEmployee);
+    await employeesRepo.AddNewEmployee(initialEmployee);
     var employeeBeforeUpdate = await context.Employees.FirstOrDefaultAsync(e => e.FirstName == "Rami" && e.LastName == "Salah");
     employeeToUpdateWith.EmployeeId = employeeBeforeUpdate!.EmployeeId;
     
     Console.WriteLine($"Employee before update: {employeeBeforeUpdate?.EmployeeId} - {employeeBeforeUpdate?.FirstName}, {employeeBeforeUpdate?.LastName}.");
 
-    await crudOperate.UpdateEmployee(employeeToUpdateWith);
+    await employeesRepo.UpdateEmployee(employeeToUpdateWith);
     var updatedEmployee = await context.Employees.FirstOrDefaultAsync(e => e.EmployeeId == employeeToUpdateWith.EmployeeId);
     
     Console.WriteLine($"Employee after updating: {updatedEmployee?.EmployeeId} - {updatedEmployee?.FirstName}, {updatedEmployee?.LastName}.");
@@ -243,7 +249,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     Console.WriteLine("Employees region clean up!");
     
-    await crudOperate.DeleteRestaurant(await context.Restaurants.FirstAsync(r => r.RestaurantId == restaurantThree!.RestaurantId));
+    await restaurantsRepo.DeleteRestaurant(await context.Restaurants.FirstAsync(r => r.RestaurantId == restaurantThree!.RestaurantId));
     Console.WriteLine("Restaurant in the employees region got Deleted!");
     
     
@@ -253,7 +259,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     // ::: Adding new Entity; Begin.
     
-    await crudOperate.AddNewRestaurant(new Restaurant()
+    await restaurantsRepo.AddNewRestaurant(new Restaurant()
     {
         Name = "Restaurant four",
         Address = "Gaza, Palestine",
@@ -263,7 +269,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
 
     var restaurantFour = await context.Restaurants.FirstOrDefaultAsync(r => r.PhoneNumber == "88776666");
     
-    await crudOperate.AddNewMenuItem(new MenuItem()
+    await menuItemsRepo.AddNewMenuItem(new MenuItem()
     {
         Name = "Coffee Machine",
         Description = "Coffee grinding machine",
@@ -279,7 +285,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     // ::: Deleting Entity; Begin.
     
-    await crudOperate.DeleteMenuItem(await context.MenuItems.FirstAsync(m => m.Name == "Coffee Machine" && m.Description == "Coffee grinding machine"));
+    await menuItemsRepo.DeleteMenuItem(await context.MenuItems.FirstAsync(m => m.Name == "Coffee Machine" && m.Description == "Coffee grinding machine"));
     var deletedMenuItem = await context.MenuItems.FirstOrDefaultAsync(m => m.Name == "Coffee Machine" && m.Description == "Coffee grinding machine");
     
     Console.WriteLine(deletedMenuItem is null ? "Menu item one got deleted successfully!" : "ERROR while deleting the menu item!");
@@ -303,13 +309,13 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         RestaurantId = restaurantFour.RestaurantId,
     };
     
-    await crudOperate.AddNewMenuItem(initialMenuItem);
+    await menuItemsRepo.AddNewMenuItem(initialMenuItem);
     var menuItemBeforeUpdate = await context.MenuItems.FirstOrDefaultAsync(m => m.Name == "Tea" && m.Description == "Green tea");
     menuItemToUpdateWith.ItemId = menuItemBeforeUpdate!.ItemId;
     
     Console.WriteLine($"Menu item before update: {menuItemBeforeUpdate?.ItemId} - {menuItemBeforeUpdate?.Name}, {menuItemBeforeUpdate?.Description}.");
 
-    await crudOperate.UpdateMenuItem(menuItemToUpdateWith);
+    await menuItemsRepo.UpdateMenuItem(menuItemToUpdateWith);
     var updatedMenuItem = await context.MenuItems.FirstOrDefaultAsync(m => m.Name == "Sugar" && m.Description == "White sugar");
     
     Console.WriteLine($"Menu item after updating: {updatedMenuItem?.ItemId} - {updatedMenuItem?.Name}, {updatedMenuItem?.Description}.");
@@ -320,7 +326,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     Console.WriteLine("Menu items region clean up!");
     
-    await crudOperate.DeleteRestaurant(await context.Restaurants.FirstAsync(r => r.RestaurantId == restaurantFour.RestaurantId));
+    await restaurantsRepo.DeleteRestaurant(await context.Restaurants.FirstAsync(r => r.RestaurantId == restaurantFour.RestaurantId));
     Console.WriteLine("Restaurant in the menu item region got deleted successfully!");
     
     
@@ -330,7 +336,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     // ::: Adding new Entity; Begin.
     
-    await crudOperate.AddNewRestaurant(new Restaurant()
+    await restaurantsRepo.AddNewRestaurant(new Restaurant()
     {
         Name = "Restaurant Five",
         Address = "Gaza, Palestine",
@@ -340,7 +346,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
 
     var restaurantFive = await context.Restaurants.FirstOrDefaultAsync(r => r.PhoneNumber == "222333444333");
     
-    await crudOperate.AddNewTable(new Table()
+    await tablesRepo.AddNewTable(new Table()
     {
         Capacity = 8,
         RestaurantId = restaurantFive!.RestaurantId,
@@ -354,7 +360,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     // ::: Deleting Entity; Begin.
     
-    await crudOperate.DeleteTable(await context.Tables.FirstAsync(t => t.TableId == table.TableId));
+    await tablesRepo.DeleteTable(await context.Tables.FirstAsync(t => t.TableId == table.TableId));
     var deletedTable = await context.Tables.FirstOrDefaultAsync(t => t.TableId == table.TableId);
     
     Console.WriteLine(deletedTable is null ? "Table one got deleted successfully!" : "ERROR while deleting the table!");
@@ -374,13 +380,13 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         RestaurantId = restaurantFive!.RestaurantId,
     };
     
-    await crudOperate.AddNewTable(initialTable);
+    await tablesRepo.AddNewTable(initialTable);
     var tableBeforeUpdate = await context.Tables.FirstOrDefaultAsync(t => t.Capacity == 5 && t.RestaurantId == restaurantFive!.RestaurantId);
     tableToUpdateWith.TableId = tableBeforeUpdate!.TableId;
     
     Console.WriteLine($"Table before update: {tableBeforeUpdate?.TableId} - {tableBeforeUpdate?.Capacity}.");
 
-    await crudOperate.UpdateTable(tableToUpdateWith);
+    await tablesRepo.UpdateTable(tableToUpdateWith);
     var updatedTable = await context.Tables.FirstOrDefaultAsync(t => t.TableId == tableToUpdateWith.TableId);
     
     Console.WriteLine($"Table after updating: {tableBeforeUpdate?.TableId} - {tableBeforeUpdate?.Capacity}.");
@@ -391,7 +397,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     Console.WriteLine("Tables region clean up!");
     
-    await crudOperate.DeleteRestaurant(await context.Restaurants.FirstAsync(r => r.RestaurantId == restaurantFive.RestaurantId));
+    await restaurantsRepo.DeleteRestaurant(await context.Restaurants.FirstAsync(r => r.RestaurantId == restaurantFive.RestaurantId));
     Console.WriteLine("Restaurant in the tables region got deleted successfully!");
     
     #endregion
@@ -400,7 +406,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     // ::: Adding new Entity; Begin.
 
-    await crudOperate.AddNewCustomer(new Customer()
+    await customersRepo.AddNewCustomer(new Customer()
     {
         FirstName = "Salama",
         LastName = "Hamada",
@@ -408,7 +414,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         PhoneNumber = "111222333",
     });
     var customerOne = await context.Customers.FirstOrDefaultAsync(c => c.Email == "salamahamada22@gamil.com");
-    await crudOperate.AddNewRestaurant(new Restaurant()
+    await restaurantsRepo.AddNewRestaurant(new Restaurant()
     {
         Name = "Restaurant Six",
         Address = "Gaza, Palestine",
@@ -416,7 +422,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         OpeningHours = new TimeOnly(),
     });
     var restaurantSix = await context.Restaurants.FirstOrDefaultAsync(r => r.PhoneNumber == "98778999");
-    await crudOperate.AddNewTable(new Table()
+    await tablesRepo.AddNewTable(new Table()
     {
         Capacity = 4,
         RestaurantId = restaurantSix!.RestaurantId,
@@ -425,7 +431,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         t.Capacity == 4 && t.RestaurantId == restaurantSix!.RestaurantId);
 
     var dateOne = DateTime.Now;
-    await crudOperate.AddNewReservation(new Reservation()
+    await reservationsRepo.AddNewReservation(new Reservation()
     {
         ReservationDate = dateOne,
         PartySize = 12,
@@ -441,7 +447,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
 
     // ::: Deleting Entity; Begin.
     
-    await crudOperate.DeleteReservation(await context.Reservations.FirstAsync(r => r.ReservationId == reservationOne.ReservationId));
+    await reservationsRepo.DeleteReservation(await context.Reservations.FirstAsync(r => r.ReservationId == reservationOne.ReservationId));
     var deletedReservation = await context.Reservations.FirstOrDefaultAsync(r => r.ReservationId == reservationOne.ReservationId);
     
     Console.WriteLine(deletedReservation is null ? "Reservation got deleted successfully!" : "ERROR while deleting the table!");
@@ -469,13 +475,13 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         TableId = tableTwo!.TableId,
     };
     
-    await crudOperate.AddNewReservation(initialReservation);
+    await reservationsRepo.AddNewReservation(initialReservation);
     var reservationBeforeUpdate = await context.Reservations.FirstOrDefaultAsync(r => r.ReservationDate == dateTwo && r.PartySize == 10);
     reservationToUpdateWith.ReservationId = reservationBeforeUpdate!.ReservationId;
     
     Console.WriteLine($"Reservation before update: {reservationBeforeUpdate?.ReservationId} - {reservationBeforeUpdate?.ReservationDate}.");
 
-    await crudOperate.UpdateReservation(reservationToUpdateWith);
+    await reservationsRepo.UpdateReservation(reservationToUpdateWith);
     var updatedReservation = await context.Reservations.FirstOrDefaultAsync(r => r.ReservationDate == dateThree && r.PartySize == 19);
     
     Console.WriteLine($"Reservation after updating: {updatedReservation?.ReservationId} - {updatedReservation?.ReservationDate}.");
@@ -486,12 +492,12 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     Console.WriteLine("Reservations region clean up!");
 
-    await crudOperate.DeleteReservation(
+    await reservationsRepo.DeleteReservation(
         await context.Reservations.FirstAsync(r => r.ReservationId == reservationToUpdateWith.ReservationId));
     Console.WriteLine("Reservation in the reservations region got deleted successfully!");
-    await crudOperate.DeleteCustomer(await context.Customers.FirstAsync(c => c.CustomerId == customerOne.CustomerId));
+    await customersRepo.DeleteCustomer(await context.Customers.FirstAsync(c => c.CustomerId == customerOne.CustomerId));
     Console.WriteLine("Customer in the reservations region got deleted successfully!");
-    await crudOperate.DeleteRestaurant(
+    await restaurantsRepo.DeleteRestaurant(
         await context.Restaurants.FirstAsync(r => r.RestaurantId == restaurantSix.RestaurantId));
     Console.WriteLine("Restaurant in the reservations region got deleted successfully!");
 
@@ -501,7 +507,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     // ::: Adding new Entity; Begin.
     
-    await crudOperate.AddNewCustomer(new Customer()
+    await customersRepo.AddNewCustomer(new Customer()
     {
         FirstName = "Jemmy",
         LastName = "Maccce",
@@ -509,7 +515,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         PhoneNumber = "223334445",
     });
     var customerTwo = await context.Customers.FirstOrDefaultAsync(c => c.Email == "jemmmmy55@gamil.com");
-    await crudOperate.AddNewRestaurant(new Restaurant()
+    await restaurantsRepo.AddNewRestaurant(new Restaurant()
     {
         Name = "Restaurant Seven",
         Address = "Gaza, Palestine",
@@ -517,7 +523,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         OpeningHours = new TimeOnly(),
     });
     var restaurantSeven = await context.Restaurants.FirstOrDefaultAsync(r => r.PhoneNumber == "12344321");
-    await crudOperate.AddNewTable(new Table()
+    await tablesRepo.AddNewTable(new Table()
     {
         Capacity = 87,
         RestaurantId = restaurantSeven!.RestaurantId,
@@ -526,7 +532,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         t.Capacity == 87 && t.RestaurantId == restaurantSeven!.RestaurantId);
 
     var dateFour = DateTime.Now.Add(new TimeSpan(22));
-    await crudOperate.AddNewReservation(new Reservation()
+    await reservationsRepo.AddNewReservation(new Reservation()
     {
         ReservationDate = dateFour,
         PartySize = 18,
@@ -535,7 +541,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         TableId = tableThree!.TableId,
     });
     var reservationTwo = await context.Reservations.FirstOrDefaultAsync(r => r.ReservationDate == dateFour && r.PartySize == 18);
-    await crudOperate.AddNewEmployee(new Employee()
+    await employeesRepo.AddNewEmployee(new Employee()
     {
         FirstName = "Ramaa",
         LastName = "Jomaa",
@@ -543,7 +549,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         RestaurantId = restaurantSeven.RestaurantId,
     });
     var employeeOne = await context.Employees.FirstOrDefaultAsync(e => e.FirstName == "Ramaa" && e.LastName == "Jomaa");
-    await crudOperate.AddNewOrder(new Order()
+    await ordersRepo.AddNewOrder(new Order()
     {
         OrderDate = dateFour,
         TotalAmount = 6677.99M,
@@ -558,7 +564,7 @@ async Task CrudOperateOnEntities(RestaurantContext context)
 
     // ::: Deleting Entity; Begin.
     
-    await crudOperate.DeleteOrder(await context.Orders.FirstAsync(o => o.OrderId == orderOne.OrderId));
+    await ordersRepo.DeleteOrder(await context.Orders.FirstAsync(o => o.OrderId == orderOne.OrderId));
     var deletedOrder = await context.Orders.FirstOrDefaultAsync(o => o.OrderId == orderOne.OrderId);
     
     Console.WriteLine(deletedOrder is null ? "Order got deleted successfully!" : "ERROR while deleting the order!");
@@ -584,13 +590,13 @@ async Task CrudOperateOnEntities(RestaurantContext context)
         EmployeeId = employeeOne!.EmployeeId,
     };
     
-    await crudOperate.AddNewOrder(initialOrder);
+    await ordersRepo.AddNewOrder(initialOrder);
     var orderBeforeUpdate = await context.Orders.FirstOrDefaultAsync(o => o.OrderDate == dateFive && o.TotalAmount == 5678.99M);
     orderToUpdateWith!.OrderId = orderBeforeUpdate!.OrderId;
     
     Console.WriteLine($"Reservation before update: {orderBeforeUpdate!.OrderId} - {orderBeforeUpdate!.OrderDate} - {orderBeforeUpdate!.TotalAmount}.");
     
-    await crudOperate.UpdateOrder(orderToUpdateWith);
+    await ordersRepo.UpdateOrder(orderToUpdateWith);
     var updatedOrder = await context.Orders.FirstOrDefaultAsync(o => o.OrderId == orderToUpdateWith.OrderId);
     
     Console.WriteLine($"Reservation after update: {updatedOrder!.OrderId} - {updatedOrder!.OrderDate} - {updatedOrder!.TotalAmount}.");
@@ -601,17 +607,17 @@ async Task CrudOperateOnEntities(RestaurantContext context)
     
     Console.WriteLine("Orders region clean up!");
 
-    await crudOperate.DeleteCustomer(await context.Customers.FirstAsync(c => c.CustomerId == customerTwo!.CustomerId));
+    await customersRepo.DeleteCustomer(await context.Customers.FirstAsync(c => c.CustomerId == customerTwo!.CustomerId));
     Console.WriteLine("Customer in the orders region got deleted successfully!");
-    await crudOperate.DeleteOrder(
+    await ordersRepo.DeleteOrder(
         await context.Orders.FirstAsync(o => o.OrderId == updatedOrder.OrderId));
     Console.WriteLine("Order in the orders region got deleted successfully!");
-    await crudOperate.DeleteReservation(await context.Reservations.FirstAsync(r => r.ReservationId == reservationTwo.ReservationId));
+    await reservationsRepo.DeleteReservation(await context.Reservations.FirstAsync(r => r.ReservationId == reservationTwo.ReservationId));
     Console.WriteLine("Reservations in the orders region got deleted successfully!");
-    await crudOperate.DeleteRestaurant(
+    await restaurantsRepo.DeleteRestaurant(
         await context.Restaurants.FirstAsync(r => r.RestaurantId == restaurantSeven!.RestaurantId));
     Console.WriteLine("Restaurant in the orders region got deleted successfully!");
-    await crudOperate.DeleteEmployee(
+    await employeesRepo.DeleteEmployee(
         await context.Employees.FirstAsync(e => e.EmployeeId == employeeOne!.EmployeeId));
     Console.WriteLine("Employee in the orders region got deleted successfully!");
 
@@ -620,15 +626,13 @@ async Task CrudOperateOnEntities(RestaurantContext context)
 
 async Task QueryEntities(RestaurantContext context)
 {
-    var entitiesQuery = new EntitiesQuery(context);
-
-    var managers = entitiesQuery.ListManagers();
-    var reservations = entitiesQuery.GetReservationsByCustomer(1);
-    var orderAndItems = entitiesQuery.ListOrdersAndMenuItems(1);
-    var menuItems = entitiesQuery.ListOrderedMenuItems(1);
-    var amountAverage = await entitiesQuery.CalculateAverageOrderAmount(1);
-    var restaurantRevenue = await entitiesQuery.CalculateRestaurantRevenue(1);
-    var customers = await entitiesQuery.CustomerReservationsWithPartySizeGreaterThan(5);
+    var managers = employeesRepo.ListManagers();
+    var reservations = reservationsRepo.GetReservationsByCustomer(1);
+    var orderAndItems = ordersRepo.ListOrdersAndMenuItems(1);
+    var menuItems = menuItemsRepo.ListOrderedMenuItems(1);
+    var amountAverage = await employeesRepo.CalculateAverageOrderAmount(1);
+    var restaurantRevenue = await restaurantsRepo.CalculateRestaurantRevenue(1);
+    var customers = await customersRepo.CustomerReservationsWithPartySizeGreaterThan(5);
     
     Console.WriteLine($"The total amount for the employee: {amountAverage}");
